Token distributor no longer joins microovn
This makes token distributor a standalone charm not part of the microovn
cluster. This means we can manage microovn entirely though the microovn
juju application.

The token distributor acts as a mirror giving the microovn charm
information it would typically get through a peer relation, this works
very efficiently and due to its simplicity allows the token distributor
charm to only be 60 loc (which makes me very happy).

There are a couple edge cases which can cause issues at the moment but
there are solid solutions for them.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>